requirements.txt
pip~=25.2
setuptools~=65.5.0
numpy==1.24.4
pandas==2.0.2
scikit-learn==1.3.2
tensorflow==2.13.0
matplotlib==3.7.5
pyyaml==6.0.3